+++ commands/demon.js
@@ -1,6 +1,6 @@
 const YZEmbed = require('../util/YZEmbed');
-const Demon = require('../util/YZDemonGenerator');
-const Util = require('../util/Util');
+const Demon = require('../util/FBLDemonGenerator');
+// const Util = require('../util/Util');
 
 module.exports = {
 	name: 'demon',
@@ -11,6 +11,28 @@ module.exports = {
 	args: false,
 	execute(args, message) {
 		const demon = new Demon();
-		console.log(demon);
+
+		const embed = new YZEmbed(
+			`${demon.form.toUpperCase()} DEMON`,
+			'*Forbidden Lands Demon*'
+		);
+
+		// Demon's attributes & Armor Rating.
+		embed.addField(
+			'Attributes',
+			`Strength: **${demon.attributes.str}**`
+				+ `\nAgility: **${demon.attributes.agi}**`
+				+ `\nWits: **${demon.attributes.wit}**`
+				+ `\nEmpathy: **${demon.attributes.emp}**`,
+			true
+		);
+		embed.addField(
+			'Body',
+			`Armor Rating: ${demon.armor}`
+				+ (demon.formEffect) ? `\nEffect: ${demon.formEffect}` : '',
+			true
+		);
+
+		return message.channel.send(embed);
 	},
 };